+++ src/binance/BinanceTools.ts
@@ -101,7 +101,7 @@ export class BinanceTools implements EdgeCurrencyTools {
   ): Promise<EdgeParsedUri> {
     const networks = { binance: true }
 
-    const { parsedUri, edgeParsedUri } = parseUriCommon({
+    const { parsedUri, edgeParsedUri } = await parseUriCommon({
       currencyInfo: this.currencyInfo,
       uri,
       networks,