+++ src/fio/FioTools.ts
@@ -150,7 +150,7 @@ export class FioTools implements EdgeCurrencyTools {
   }
 
   async parseUri(uri: string): Promise<EdgeParsedUri> {
-    const { edgeParsedUri } = parseUriCommon({
+    const { edgeParsedUri } = await parseUriCommon({
       currencyInfo: this.currencyInfo,
       uri,
       networks: { fio: true },