+++ src/hedera/HederaTools.ts
@@ -143,7 +143,7 @@ export class HederaTools implements EdgeCurrencyTools {
     const {
       edgeParsedUri,
       edgeParsedUri: { publicAddress }
-    } = parseUriCommon({
+    } = await parseUriCommon({
       currencyInfo: this.currencyInfo,
       uri,
       networks: { [`${pluginId}`]: true },