+++ src/piratechain/PiratechainTools.ts
@@ -162,7 +162,7 @@ export class PiratechainTools implements EdgeCurrencyTools {
     const {
       edgeParsedUri,
       edgeParsedUri: { publicAddress }
-    } = parseUriCommon({
+    } = await parseUriCommon({
       currencyInfo: this.currencyInfo,
       uri,
       networks,