+++ src/tron/TronTools.ts
@@ -133,7 +133,7 @@ export class TronTools implements EdgeCurrencyTools {
     const networks = { [this.currencyInfo.pluginId]: true, 'usdt-trc20': true }
     const { smartPayPublicAddress, smartPayUserId } = this.initOptions
 
-    const { parsedUri, edgeParsedUri } = parseUriCommon({
+    const { parsedUri, edgeParsedUri } = await parseUriCommon({
       currencyInfo: this.currencyInfo,
       uri,
       networks,