+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added: Add optional `testPrivateKeys` function to `parseUriCommon`
 - changed: (Zcash) Updated address explorer
 - changed: (Tron) Special case the `usdt-trc20` uri prefix
 - fixed: Prevent fatal error reporting for missing txlist json file